+++ tests/php-unit-tests/composer.json
@@ -2,5 +2,11 @@
   "require-dev": {
     "phpunit/phpunit": "^8.5.23",
     "sempro/phpunit-pretty-print": "^1.4"
+  },
+  "autoload": {
+    "psr-4": {
+      "Combodo\\iTop\\Test\\UnitTest\\": "src/BaseTestCase/",
+      "Combodo\\iTop\\Test\\UnitTest\\Service\\": "src/Service/"
+    }
   }
 }
