+++ tests/php-unit-tests/unittestautoload.php
@@ -1,10 +1,7 @@
 <?php
 
 // Main autoload, this is the one to use in the PHPUnit configuration
-// It is customized to include both
-// - Vendors
+//
+// It was previously used to include both the vendor autoloader and our custom base test case classes, but these are now autoloaded from ./src/BasetestCase
+// This file should then no longer be necessary, but we have to keep it until projects / branches / modules have been corrected.
 require_once 'vendor/autoload.php';
-// - Custom test case PHP classes
-require_once 'ItopTestCase.php';
-require_once 'ItopDataTestCase.php';
-require_once 'ItopCustomDatamodelTestCase.php';
